Feat: Upgrade oracle deviation technology inside the Oracle Manager to work by asset instead of equally across all oracles. This deviation configuration also factors in underlying price feed deviation thresholds to prevent human error on configuration of restrictive deviation bound values.
Chore: Added a ton of new documentation to Oracle Manager and cleaned up naming conventions. We previous referred to adaptors as "feeds" inside the oracle manager when it hooks up to corresponding oracle adaptors, and then the adaptors hook up to the aggregators/feeds. 

Fix: Fixed bug related to NotifyFeedRemoval reverting when called with no adaptor dependencies set.